perf(dialog): reduce change detection cycles
This PR reduces change detection cycles for the dialog component by replacing
`HostListener` with `Renderer.listen` outside of the zone. The change detection should
be run only in specific cases, e.g. if the `esc` button has been clicked or the document
has been clicked and the target contains `dialog` class, but it shouldn't be run on each
button press and click event.
Note that the `HostListener` wraps the actual listener under the hood into the internal Angular
function which runs `markDirty()` before running the actual listener (the decorated class method).

Signed-off-by: Artur Androsovych <arthurandrosovich@gmail.com>