perf(drawer): reduce change detection cycles
This PR reduces change detection cycles for the drawer component by replacing
`HostListener` with `Renderer.listen` outside of the zone. The change detection should
be run only if the `Escape` button is clicked.
Note that the `HostListener` wraps the actual listener under the hood into the internal Angular
function which runs `markDirty()` before running the actual listener (the decorated class method).

The `DrawerService` has been updated to run the change detection only when the parent container is clicked.

Signed-off-by: Artur Androsovych <arthurandrosovich@gmail.com>